% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_article_at_headings.R
\name{cut_articles_at_headings}
\alias{cut_articles_at_headings}
\title{Cut article text at specified headings}
\usage{
cut_articles_at_headings(articles, sections_to_remove)
}
\arguments{
\item{articles}{A character vector of articles to be processed}

\item{sections_to_remove}{A character vector of headings at which to cut the articles (removing all subsequent text)}
}
\value{
A character vector of processed articles containing all text preceding headings
}
\description{
Cut article text at specified headings
}
