% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_category_members.R
\name{get_category_members}
\alias{get_category_members}
\title{Get Category Members}
\usage{
get_category_members(category, lang = "en", namespace = 14, verbose = TRUE)
}
\arguments{
\item{category}{The name of the category to retrieve members for.}

\item{lang}{The language code for the Wikipedia site to query (default is
"en").}

\item{namespace}{The namespace to retrieve members for (0 for pages, 14 for
categories; default is 14).}
}
\value{
A dataframe with one row for each category member, containing
  information such as title and namespace.
}
\description{
This function retrieves a list of pages or subcategories within a given
category from the Wikipedia MediaWiki API.
}
\examples{
get_category_members("Animals")
get_category_members("Moral philosophers")

}
