% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_category_members_matrix.R
\name{get_category_members_matrix}
\alias{get_category_members_matrix}
\title{Retrieve Category Members and Create a Binary Sparse Matrix}
\usage{
get_category_members_matrix(
  categories,
  depth = 1,
  lang = "en",
  namespace = 14,
  verbose = TRUE
)
}
\arguments{
\item{categories}{a list of categories to retrieve members for}

\item{depth}{the number of levels to retrieve members for (default is 1)}

\item{lang}{the language to retrieve members in (default is "en")}

\item{namespace}{the namespace to retrieve members from (default is 14,
categories)}

\item{verbose}{logical indicating whether or not to print progress updates
(default is TRUE)}
}
\value{
a binary sparse matrix denoting whether the entity in the column
  belongs to the entity in the row
}
\description{
Takes a vector of categories and retrieves all of the members (pages and
subcategories) associated with each category, returning a binary sparse
matrix denoting whether the entity in the column (a category or page) belongs
to the entity in the row (a category). This function can also be trace the
category hierarchy to a specified depth (e.g., retrieving members of
members).
}
\examples{
categories = c("Category:Animals", "Category:Plants")
get_category_members_matrix(categories)
}
