% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_headings.R
\name{get_headings}
\alias{get_headings}
\title{Extract Headings from Wikipedia Text}
\usage{
get_headings(text, heading_pattern = "\\n *={2} *[^=].+ *={2} *\\n")
}
\arguments{
\item{text}{A character string of Wikipedia text.}

\item{heading_pattern}{A regular expression pattern used to match headings. Default is "\n *={2} *[^=].+ *={2} *\n" which matches level 1 headings.}
}
\value{
A character vector of headings, with the formatting ("==") removed.
}
\description{
This function extracts headings from a Wikipedia text using a regular expression pattern.
}
\examples{
text = "== Head1 ==\nText1\n== Head2 ==\nText2"
get_headings(text)
# Output: "Head1" "Head2"
}
