% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_links.R
\name{get_links}
\alias{get_links}
\title{Retrieve links from a Wikipedia page}
\usage{
get_links(page, direction = "outgoing", lang = "en", namespaces = c(0))
}
\arguments{
\item{page}{A title (string) of the Wikipedia page to retrieve links from}

\item{direction}{A string indicating the type of links to retrieve. Must be
either "incoming" (links pointing to the specified page) or "outgoing"
(links on the specified page). Default is "outgoing".}

\item{lang}{The language code of the Wikipedia to retrieve links from.
Default is "en" (English).}

\item{namespaces}{An integer vector of namespaces to filter links by. Default
is c(0) (main namespace).}
}
\value{
A data frame containing the title, namespace, and page ID of all
  retrieved links.
}
\description{
Retrieve links from a Wikipedia page
}
\examples{
get_links("Psychology")
get_links("Drop bear", direction = "outgoing")

}
