% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wikipedia_article.R
\name{get_wikipedia_article}
\alias{get_wikipedia_article}
\title{Retrieve Wikipedia Article Text}
\usage{
get_wikipedia_article(title, lang = "en", verbose = TRUE)
}
\arguments{
\item{title}{The title of the Wikipedia article to retrieve.}

\item{lang}{The language of the Wikipedia article to retrieve. Default is
"en" for English.}

\item{verbose}{logical indicating whether or not to print progress updates
(default is TRUE)}
}
\value{
A character string containing the text of the Wikipedia article.
}
\description{
This function retrieves the text of a Wikipedia article based on the provided
title and language.
}
\examples{
get_wikipedia_article("R programming language")
get_wikipedia_article("Programmation R", lang = "fr")

}
