% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heading_frequency.R
\name{plot_heading_frequency}
\alias{plot_heading_frequency}
\title{Plot Heading Frequency}
\usage{
plot_heading_frequency(
  heading_freq,
  to_file = TRUE,
  n = 25,
  path = "output/heading_frequency.pdf"
)
}
\arguments{
\item{heading_freq}{a numeric vector representing the frequency of headings.}

\item{to_file}{a logical indicating whether the plot should be saved to a file. Defaults to TRUE.}

\item{n}{an integer indicating the number of headings to be plotted. Defaults to 25.}

\item{path}{A string indicating the path and file name for saving the plot. Default is "output/heading_frequency.pdf"}
}
\value{
Plots the heading frequency, if to_file is TRUE, saves the plot as a PDF.
}
\description{
Plot Heading Frequency
}
