% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_wikipedia_text.R
\name{split_wikipedia_text}
\alias{split_wikipedia_text}
\title{Split Wikipedia Text by Headings}
\usage{
split_wikipedia_text(text)
}
\arguments{
\item{text}{A character string of Wikipedia text.}
}
\value{
A named character vector, with the names corresponding to the
  headings in the input text. The first element of the list is named "Lead"
  and contains the text before the first heading.
}
\description{
This function splits Wikipedia text into segments based on level 1 headings.
}
